requirements.txt
python-etcd
daemonize
jsonschema
